use addressindex to check for MN collateral not txindex
Associated with issue #1734 in that txindex was being forced upon nodes to ensure that masternode collateral was checked properly. Pruned nodes would not work without txindex because they wouldn't have the collateral transaction history to be able to verify old masternodes.

However by using addressindex which is pruning compatible and a lot lighter than txindex we can verify masternode collateral without txindex (must set addressindex instead).

Also not that dual calls of GetTransaction in CheckOutpoint can be easily avoided by returning the height from IsInputAssociatedWithPubkey since addressindex returns height when UTXO was created.

Other calls to GetTransaction stay as is, one is in CGovernanceObject::IsCollateralValid where it will do a slow lookup but also assume that collateral objects are fairly new and new nodes will generally keep a months worth of data around to be able to verify valid proposals.

The other call is in CInstantSend::ResolveConflicts and again there is an assumption that pruning nodes would atleast have enough data around to lock and verify an instantx.