Refactor and optimize IsInputAssociatedWithPubkey
Remove the extra GetTransaction call by passing height back from IsInputAssociatedWithPubkey function and using it in CheckOutpoint. Also remove GetTransaction call in IsInputAssociatedWithPubkey in favor of GetUTXOCoins which is more efficient and does the same thing. The only difference is that GetUTXOCoins will fail if the collateral has moved whereas GetTransaction will not. I hope that is a safe assumption as if collateral is moved I would like to think that this function should fail anyway (seems to be only called when a masternode broadcast has occured)