Fix CreateNewBlock_validitymine tests
The test increments the chain tip height without actually putting a block
on-top of the chain. This was always invalid but is never really checked
so the test passed. DIP3 introduced some consistency checks that also check
for the expected height of a new block. This commit allows to skip block
validity tests when creating a new block template.