use CLIENT_VERSION as default daemon version instead of hardcoded value
This magic constant was not being used to set the daemon version inside of CMasternodePing's constructor with outpoint. It caused me a 2 hour headache in Syscoin where I figured out the hash of the mnp was different from sender and reciever side and thus signature verification failed because the daemon version's were inadvertently different because on verification
it overrides the daemon with the local client version where as on the sender it uses the DEFAULT_DAEMON_VERSION and they just so happened to be different for me.