Add protx infoall command
This is used to list the full DMN list without writing some shell script to do
so (and in one single RPC command). This is useful because when I am trying to
issue a protx service update from the MN itself (with the operator key), there
is no easy way of finding the protx hash. But this allows an operator to search
thru the list for the IP address.

Another nice feature would be a filter to this same command, but I haven't
implemented it yet.