Don't use boost::thread_specific_ptr in LogAcceptCategory
Using valgrind memcheck shows that this causes invalid "free" calls on exit.
I'm not sure what the reason is and it might actually be a false positive.
However, this stops us from using valgrind memcheck and valgrind massif,
so we should fix this issue by using the C++11 thread_local storage
specifier.