Fix relaying of governance objects when collateral is only IS/CL locked
A bug in CGovernanceObject::IsCollateralValid caused governance objects
to be rejected when the collateral was IS locked and not fully confirmed
on-chain. The intention was to actually accept such objects and relay them.

This changes IsCollateralValid to accept such collaterals but also gives
the caller the required info to postpone relaying of such objects for older
peers.