Remove dataChanged call for ToAddress
This call is redundant since a long time already. It comes from a time
where there was no ToAddress column, but only a Description column which
also contained status information for coinbases. This is gone for a long
time already, so the ToAddress column has nothing included which might
change when confirmations change.

For the history:
34fa178243d462fa8ed8978227626bcd0c622d82 splitted the Description column
into Type and ToAddress
e59924680340427b658d1bb37dc38028f6fa6317 removed last traces of status from
the ToAddress column