Fix undefined behaviour in stacktrace printing.
Not returning a value from a function with non-void return type
is undefined behaviour in C++ (except for main() function).

Fortunately, this was causing a crash for multiwallet functional test
when building with GCC 9.2 with default optimisation flags for building
Fedora packages, so I was able to investigate, find this bug and fix it.

But even if it was not causing a crash and just returning some random
garbage in eax register, it would lead to incorrect behaviour:
non-zero return from a handler causes dl_iterate_phdr() to stop iteration.

Signed-off-by: Oleg Girko <ol@infoserver.lv>