refactor: introduce MAKE_MSG macro for compile time check to ensure the p2p message name is short enough
I would like this ofc to not be a macro, but to be as non-invasive as possible, I think a macro is best. Arguably these should all be converted to string_view and have a better abstraction layer over it (kinda like we do with sporks) as opposed to needing to place these names in multiple files.