refactor: use `WalletLoader` to access wallets in `WalletInit` funcs
This is required because we're going to get deglobalize wallet variables
in an upcoming commit, which requires us to hold `WalletContext` in
order to iterate through all wallets. So, we rely on the interface to
help us do that.

Also, let's fold in the `WalletInit::InitCoinJoinSettings` calls into
`interfaces::CoinJoin::Loader` for simplicity's sake.