rpc: allow `addresses` to report data from legacy platform port fields
MnNetInfo doesn't actually store this information so we are pulling the
data from platform{HTTP,P2P}Port to make it appear like it does. This
is to ensure reporting parity with ExtNetInfo, which stores platform
fields and should allow us to deprecate the platform{HTTP,P2P}Port when
we need to.