perf: reduce timing intervals in CSigSharesManager for improved responsiveness
Adjusted the timing intervals in the WorkThreadMain function of CSigSharesManager to 10 milliseconds for both message sending and work interruption checks. This change enhances the responsiveness of the signing shares manager by allowing more frequent processing of pending signature shares and message sending.

This results in ~33% latency improvements in a contrived local latency functional test / benchmark from ~500ms to ~333ms