feat(qt): Add flag to allow duplicate recipient addresses in GUI
This enables users of the Qt wallet to send to the same address in a single
transaction. This is useful for setting up multiple MN collateral transactions,
but probably only for advanced users and developers.

Usage: ./qt/dash-qt -allowduplicaterecipients

I'm ok if there's no desire to merge this. I just thought it was a nice feature
to send all collateral in a single TX instead of having to send multiple. I
presume this is mostly just useful for testnet, as a hardware wallet would be
preferred for mainnet, but it may also work to create an unsigned TX which
could be signed by a HW wallet w/the right UI.